A and B
Completed a and b
Moved b to own file.
